+++ src/openthread_rcp/CHANGELOG.md
@@ -1,3 +1,9 @@
+# SL-OPENTHREAD/3.1.0.0_GitHub-fb274efe6
+This beta release is built with OpenThread 3.1.0 from Simplicity SDK 2026.6.0.
+
+# SL-OPENTHREAD/3.0.2.0_GitHub-61e43cffb
+This beta release is built with OpenThread 3.0.2 from Simplicity SDK 2025.12.3.
+
 # SL-OPENTHREAD/2.7.2.0_GitHub-fb0446f53
 This beta release is built with Simplicity SDK 2025.6.2.